# lib/db: Custom Bloom filter implementation