--- a/lib/db/bloomfilter.go
+++ b/lib/db/bloomfilter.go
@@ -1,3 +1,9 @@
+// Copyright (C) 2020 The Syncthing Authors.
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
 package db
 
 import (