@calmh 2020-04-02
Sweet. This looks like a neat and fast implementation, and I assume you wrote it custom for the purpose given the copyright. Thing is though, after looking at it for a few minutes and being only familiar with the high level details of how a bloom filter works, I can't claim to understand every bitshift in it. Given that and that open source is generally free as in a puppy (adopt it and maintain for life), I'm not 100% sure I want to become a maintainer for a bespoke bloom filter implementation.

How about a github.com/greatroar/bloom package, with a real person copyright, license, readme, and releases, and then a PR to use that as a dependency? You might say that's precisely the same thing, and to some extent it is, but it delineates maintenance responsibilities slightly differently.